Introduction

ClimateBase is a Julia package offering basic functionality for analyzing data that are typically in the form used by climate sciences. These data are dimensional & spatiotemporal but the corresponding dimensions all need special handling. For example the most common dimensions are longitude, latitude and time.

longitude is by definition a periodic dimension

latitude is a linear dimension. However because the coordinate system most often used in climate sciences is a grid of longitude × latitude (in equal degrees) the area element of space depends on latitude and this needs to be taken into account.

time is a linear dimension in principle, but its values are <: AbstractDateTime instead of <: Real. The human calendar (where these values come from) is periodic but each period may not correspond to the same physical time, and this also needs to be taken into account.

ClimateBase is structured to deal with these intricacies, and in addition offer several functionalities commonly used, and sought after, by climate scientists. It also serves as the base building block for ClimateTools, which offers more advanced functionalities.

At the moment the focus of ClimateBase is not on operating on data on disk. It is designed for in-memory climate data exploration and manipulation.

Installation

This package is registered and you can install it with

using Pkg; Pkg.add("ClimateBase")

Make sure your installed version coincides with the one in this docs (see bottom left corner of this page).

ClimArray: the core data structure

This project treats "climate data" as an instance of ClimArray. At the moment ClimArray is a subtype of DimensionalArray from DimensionalData.jl. A brief introduction to DimensionalData.jl is copied here from its docs, because basic handling of a ClimArray comes from DimensionalData.jl. DimensionalData.jl allows to dimensionally-index data by their values.

E.g. you can create an array with

using ClimateBase, Dates
Time = ClimateBase.Ti # `Time` is more intuitive than `Ti`
lats = -90:5:90
lons = 0:10:359
t = Date(2000, 3, 15):Month(1):Date(2020, 3, 15)
# Here we wrap all dimension data into proper dimensions:
dimensions = (Lon(lons), Lat(lats), Time(t))
# where `Lon, Lat, Time` are `Dimension`s exported by ClimateBase
# combining the array data with dimensions makes a `ClimArray`:
data = rand(36, 37, 241) # same size as `dimensions`
A = ClimArray(data, dimensions)

ClimArray with dimensions:
  Lon (Longitude): 0:10:350 Sampled: Ordered Regular Points,
  Lat (Latitude): -90:5:90 Sampled: Ordered Regular Points,
  Ti (Time): Dates.Date("2000-03-15"):Dates.Month(1):Dates.Date("2020-03-15")
    Sampled: Ordered Regular Points
and data: 36×37×241 Array{Float64, 3}
[:, :, 1]
 0.0398387  0.648253  0.629547   0.795659  …  0.898269   0.264741  0.123295
 0.823571   0.859515  0.417002   0.260969     0.12105    0.919887  0.143839
 0.122334   0.240469  0.514432   0.519491     0.956794   0.27006   0.933766
 0.150566   0.52845   0.0233252  0.411222     0.312592   0.767053  0.390648
 ⋮                                         ⋱             ⋮         
 0.128144   0.331058  0.845882   0.351735     0.899753   0.895495  0.51027
 0.0914629  0.56654   0.972868   0.723017     0.0265146  0.189499  0.0915125
 0.248601   0.181444  0.812635   0.556306     0.968634   0.353213  0.437088
 0.545978   0.854938  0.819252   0.608276  …  0.0716127  0.964069  0.225364
[and 240 more slices...]

You can then select a specific time slice at Date(2011,5,15) and a longitude interval between 0 and 30 degrees like so:

B = A[Lon(Between(0, 30)), Time(At(Date(2011,5,15)))]

ClimArray with dimensions:
  Lon (Longitude): 0:10:30 Sampled: Ordered Regular Points,
  Lat (Latitude): -90:5:90 Sampled: Ordered Regular Points
and data: 4×37 Matrix{Float64}
 0.143788  0.805047  0.528913  0.695244  …  0.772316  0.831868  0.96128
 0.954591  0.567639  0.28287   0.184764     0.565341  0.905315  0.160215
 0.81084   0.222423  0.028815  0.526036     0.468546  0.779156  0.331323
 0.396081  0.670499  0.26567   0.198964     0.852961  0.800284  0.234405

With ClimArray you can use convenience, physically-inspired functions that do automatic (and correct) weighting. For example the latitudinal mean of B is simply

C = latmean(B)

ClimArray with dimensions:
  Lon (Longitude): 0:10:30 Sampled: Ordered Regular Points
and data: 4-element Vector{Float64}
[0.557209, 0.435859, 0.518414, 0.525704]

where in this averaging process each data point is weighted by the cosine of its latitude.

Making a ClimArray

You can create a ClimArray yourself, or you can load data from an .nc file with CF-conventions, see NetCDF IO.

ClimateBase.ClimArray — Method

ClimArray(A::Array, dims::Tuple; name = "", attrib = nothing)

ClimArray is a structure that contains numerical array data bundled with dimensional information, a name and an attrib field (typically a dictionary) that holds general attributes. You can think of ClimArray as a in-memory representation of a CFVariable.

At the moment, a ClimArray is using DimensionalArray from DimensionalData.jl, and all basic handling of ClimArray is offered by DimensionalData (see below).

ClimArray is created by passing in standard array data A and a tuple of dimensions dims. See ncread to automatically create a ClimArray from a .nc file.

Example

using ClimateBase, Dates
Time = ClimateBase.Ti # more intuitive name for time dimension
lats = -90:5:90
lons = 0:10:359
t = Date(2000, 3, 15):Month(1):Date(2020, 3, 15)
# dimensional information:
dimensions = (Lon(lons), Lat(lats), Time(t))
data = rand(36, 37, 241) # numeric data
A = ClimArray(data, dimensions)

It is strongly recommended to use the dimensions we export (because we dispatch on them and use their information):

for D in ClimateBase.STANDARD_DIMS
    println(D, " (full name = $(DimensionalData.name(D)))")
end

Lon (full name = Longitude)
Lat (full name = Latitude)
Ti (full name = Time)
Hei (full name = Height)
Pre (full name = Pressure)
Coord (full name = )

We explicitly assume that Lon, Lat are measured in degrees and not radians or meters (extremely important for spatial averaging processes).

Crash-course to DimensionalData.jl

DimensionalData — Module

DimensionalData

DimensionalData.jl provides tools and abstractions for working with datasets that have named dimensions, and optionally a lookup index. It's a pluggable, generalised version of AxisArrays.jl with a cleaner syntax, and additional functionality found in NamedDims.jl. It has similar goals to pythons xarray, and is primarily written for use with spatial data in GeoData.jl.

Broadcasting and most Base methods maintain and sync dimension context.

DimensionalData.jl also implements:

comprehensive plot recipes for Plots.jl.

a Tables.jl interface with DimTable

multi-layered DimStacks that can be indexed together, and have base methods applied to all layers.

the Adapt.jl interface for use on GPUs, even as GPU kernel arguments.

traits for handling a wide range of spatial data types accurately.

Dimensions

Dimensions are wrapper types. They hold the lookup index, details about the grid, and other metadata. They are also used to index into the array. X, Y, Z and Ti are the exported defaults. A generalised Dim type is available to use arbitrary symbols to name dimensions. Custom dimension types can also be defined using the @dim macro.

Dimensions can be used to construct arrays in rand, ones, zeros and fill with either a range for a lookup index or a number for the dimension length:

julia> using DimensionalData

julia> A = rand(X(1:40), Y(50))
40×50 DimArray{Float64,2} with dimensions:
  X: 1:40 (Sampled - Ordered Regular Points)
  Y
 0.929006   0.116946  0.750017  …  0.172604  0.678835   0.495294
 0.0550038  0.100739  0.427026     0.778067  0.309657   0.831754
 ⋮                              ⋱
 0.647768   0.965682  0.049315     0.220338  0.0326206  0.36705
 0.851769   0.164914  0.555637     0.771508  0.964596   0.30265

We can also use dim wrappers for indexing, so that the dimension order in the underlying array does not need to be known:

julia> A[Y(1), X(1:10)]
10-element DimArray{Float64,1} with dimensions:
  X: 1:10 (Sampled - Ordered Regular Points)
and reference dimensions: Y(1) 
 0.929006
 0.0550038
 0.641773
 ⋮
 0.846251
 0.506362
 0.0492866

And this has no runtime cost:

julia> A = ones(X(3), Y(3))
3×3 DimArray{Float64,2} with dimensions: X, Y
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> @btime $A[X(1), Y(2)]
  1.077 ns (0 allocations: 0 bytes)
1.0

julia> @btime parent($A)[1, 2]
  1.078 ns (0 allocations: 0 bytes)
1.0

Dims can be used in X=val form:

julia> A = rand(X(40), Y(50))
40×50 DimArray{Float64,2} with dimensions: X, Y
 0.377696  0.105445  0.543156  …  0.844973  0.163758  0.849367
 ⋮                             ⋱
 0.431454  0.108927  0.137541     0.531587  0.592512  0.598927

julia> A[Y=3]
40-element DimArray{Float64,1} with dimensions: X
and reference dimensions: Y(3)
 0.543156
 ⋮
 0.137541

For views:

julia> view(A, Y(), X(1:5))
5×50 DimArray{Float64,2} with dimensions: X, Y
 0.377696  0.105445  0.543156  …  0.844973  0.163758  0.849367
 ⋮                             ⋱
 0.875279  0.133032  0.925045     0.156768  0.736917  0.444683

And for specifying dimension number in all Base and Statistics functions that have a dims argument:

julia> using Statistics

julia> A = rand(X(3), Y(4), Ti(5));

julia> mean(A; dims=Ti)
3×4×1 DimArray{Float64,3} with dimensions: X, Y, Ti (Time)
[:, :, 1]
 0.168058  0.52353   0.563065  0.347025
 0.472786  0.395884  0.307846  0.518926
 0.365028  0.381367  0.423553  0.369339

You can also use symbols to create Dim{X} dimensions. We can't use the rand method directly with Symbols, so insteadd use the regular DimArray constructor:

julia> A = DimArray(rand(10, 20, 30), (:a, :b, :c));

julia> A[a=2:5, c=9]

4×20 DimArray{Float64,2} with dimensions: Dim{:a}, Dim{:b}
and reference dimensions: Dim{:c}(9)
 0.134354  0.581673  0.422615  …  0.410222   0.687915  0.753441
 0.573664  0.547341  0.835962     0.0353398  0.794341  0.490831
 0.166643  0.133217  0.879084     0.695685   0.956644  0.698638
 0.325034  0.147461  0.149673     0.560843   0.889962  0.75733

Selectors

Selectors find indices in the lookup index, for each dimension:

At(x): get the index exactly matching the passed in value(s)

Near(x): get the closest index to the passed in value(s)

Where(f::Function): filter the array axis by a function of the dimension index values.

Between(a, b): get all indices between two values, excluding the high value.

Contains(x): get indices where the value x falls within the interval, exluding the upper value. Only used for Sampled Intervals, for Points, use At.

(Between and Contains exlude the upper boundary so that adjacent selections never contain the same index)

Selectors can be used in getindex, setindex! and view to select indices matching the passed in value(s)

We can use selectors inside dim wrappers:

julia> using Dates

julia> timespan = DateTime(2001,1):Month(1):DateTime(2001,12)
DateTime("2001-01-01T00:00:00"):Month(1):DateTime("2001-12-01T00:00:00")

julia> A = DimArray(rand(12,10), (Ti(timespan), X(10:10:100)))
12×10 DimArray{Float64,2} with dimensions:
  Ti (Time): DateTime("2001-01-01T00:00:00"):Month(1):DateTime("2001-12-01T00:00:00") (Sampled - Ordered Regular Points)
  X: 10:10:100 (Sampled - Ordered Regular Points)
 0.14106   0.476176  0.311356  0.454908  …  0.464364  0.973193  0.535004
 ⋮                                       ⋱
 0.522759  0.390414  0.797637  0.686718     0.901123  0.704603  0.0740788

julia> @btime A[X(Near(35)), Ti(At(DateTime(2001,5)))]
0.3133109280208961

Without dim wrappers selectors must be in the right order:

using Unitful

julia> A = rand(X((1:10:100)u"m"), Ti((1:5:100)u"s"));

julia> A[Between(10.5u"m", 50.5u"m"), Near(23u"s")]
4-element DimArray{Float64,1} with dimensions:
  X: (11:10:41) m (Sampled - Ordered Regular Points)
and reference dimensions:
  Ti(21 s) (Time): 21 s (Sampled - Ordered Regular Points)
 0.584028
 ⋮
 0.716715

Methods where dims can be used containing indices or Selectors

getindex, setindex! view

Methods where dims, dim types, or Symbols can be used to indicate the array dimension:

size, axes, firstindex, lastindex

cat, reverse, dropdims

reduce, mapreduce

sum, prod, maximum, minimum,

mean, median, extrema, std, var, cor, cov

permutedims, adjoint, transpose, Transpose

mapslices, eachslice

Methods where dims can be used to construct DimArrays:

fill, ones, zeros, rand

Warnings

Indexing with unordered or reverse order arrays has undefined behaviour. It will trash the dimension index, break searchsorted and nothing will make sense any more. So do it at you own risk. However, indexing with sorted vectors of Int can be useful. So it's allowed. But it will still do strange things to your interval sizes if the dimension span is Irregular.

Alternate Packages

There are a lot of similar Julia packages in this space. AxisArrays.jl, NamedDims.jl, NamedArrays.jl are registered alternative that each cover some of the functionality provided by DimensionalData.jl. DimensionalData.jl should be able to replicate most of their syntax and functionality.

AxisKeys.jl and AbstractIndices.jl are some other interesting developments. For more detail on why there are so many similar options and where things are headed, read this thread.

Available selectors

TODO