Introduction

ClimateBase is a Julia package offering basic functionality for analyzing data that are typically in the form used by climate sciences. These data are dimensional & spatiotemporal but the corresponding dimensions all need special handling. For example the most common dimensions are longitude, latitude and time.

longitude is by definition a periodic dimension

latitude is a linear dimension. However because the coordinate system most often used in climate sciences is a grid of longitude × latitude (in equal degrees) the area element of space depends on latitude and this needs to be taken into account.

time is a linear dimension in principle, but its values are <: AbstractDateTime instead of <: Real. The human calendar (where these values come from) is periodic but each period may not correspond to the same physical time, and this also needs to be taken into account.

ClimateBase is structured to deal with these intricacies, and in addition offer several functionalities commonly used, and sought after, by climate scientists. It also serves as the base building block for ClimateTools, which offers more advanced functionalities.

At the moment the focus of ClimateBase is not on operating on data on disk. It is designed for in-memory climate data exploration and manipulation.

Installation

This package is registered and you can install it with

using Pkg; Pkg.add("ClimateBase")

Make sure your installed version coincides with the one in this docs (see bottom left corner of this page).

ClimArray: the core data structure

This project treats "climate data" as an instance of ClimArray. At the moment ClimArray is a subtype of DimArray from DimensionalData.jl. A brief introduction to DimensionalData.jl is copied here from its docs, because basic handling of a ClimArray comes from DimensionalData.jl. DimensionalData.jl allows to dimensionally-index data by their values.

E.g. you can create an array with

using ClimateBase, Dates
Time = ClimateBase.Ti # `Time` is more intuitive than `Ti`
lats = -90:5:90
lons = 0:10:359
t = Date(2000, 3, 15):Month(1):Date(2020, 3, 15)
# Here we wrap all dimension data into proper dimensions:
dimensions = (Lon(lons), Lat(lats), Time(t))
# where `Lon, Lat, Time` are `Dimension`s exported by ClimateBase
# combining the array data with dimensions makes a `ClimArray`:
data = rand(36, 37, 241) # same size as `dimensions`
A = ClimArray(data, dimensions)

ClimArray with dimensions: 
  Lon Sampled 0:10:350 ForwardOrdered Regular Points,
  Lat Sampled -90:5:90 ForwardOrdered Regular Points,
  Ti Sampled Dates.Date("2000-03-15"):Dates.Month(1):Dates.Date("2020-03-15") ForwardOrdered Regular Points
and data: 36×37×241 Array{Float64, 3}
[:, :, 1]
 0.88811   0.474845  0.485183  0.0604275  …  0.0435835  0.528158  0.0037455
 0.459623  0.954147  0.476794  0.433398      0.855685   0.913875  0.774491
 0.531902  0.108424  0.96233   0.285074      0.259575   0.356773  0.103045
 0.548468  0.695225  0.79872   0.300507      0.54987    0.826006  0.584341
 ⋮                                        ⋱             ⋮         
 0.992593  0.83722   0.62236   0.34106       0.435797   0.926446  0.318077
 0.495742  0.60361   0.218905  0.0790675     0.515324   0.058565  0.577476
 0.4717    0.211474  0.996879  0.133059      0.463309   0.155407  0.0518041
 0.511172  0.352218  0.827124  0.437741   …  0.562199   0.927975  0.749465
[and 240 more slices...]

You can then select a specific time slice at Date(2011,5,15) and a longitude interval between 0 and 30 degrees like so:

B = A[Lon(Between(0, 30)), Time(At(Date(2011,5,15)))]

ClimArray with dimensions: 
  Lon Sampled 0:10:30 ForwardOrdered Regular Points,
  Lat Sampled -90:5:90 ForwardOrdered Regular Points
and data: 4×37 Matrix{Float64}
 0.498958  0.819531  0.655161  0.738314  …  0.131651  0.474843  0.979349
 0.889101  0.726213  0.108981  0.557819     0.825845  0.328824  0.393178
 0.646765  0.462987  0.674986  0.570122     0.658705  0.333713  0.1118
 0.622484  0.12106   0.294883  0.135153     0.636706  0.115207  0.851184

With ClimArray you can use convenience, physically-inspired functions that do automatic (and correct) weighting. For example the latitudinal mean of B is simply

C = latmean(B)

ClimArray with dimensions: 
  Lon Sampled 0:10:30 ForwardOrdered Regular Points
and data: 4-element Vector{Float64}
[0.545703, 0.462818, 0.407419, 0.486559]

where in this averaging process each data point is weighted by the cosine of its latitude.

Making a ClimArray

You can create a ClimArray yourself, or you can load data from an .nc file with CF-conventions, see NetCDF IO.

ClimateBase.ClimArray — Method

ClimArray(A::Array, dims::Tuple; name = "", attrib = nothing)

ClimArray is a structure that contains numerical array data bundled with dimensional information, a name and an attrib field (typically a dictionary) that holds general attributes. You can think of ClimArray as a in-memory representation of a CFVariable.

At the moment, a ClimArray is using DimArray from DimensionalData.jl, and all basic handling of ClimArray is offered by DimensionalData (see below).

ClimArray is created by passing in standard array data A and a tuple of dimensions dims. See ncread to automatically create a ClimArray from a .nc file. For obtaining the raw numeric values of a ClimArray or any of its dimensions, use the function gnv.

Example

using ClimateBase, Dates
Time = ClimateBase.Ti # more intuitive name for time dimension
lats = -90:5:90
lons = 0:10:359
t = Date(2000, 3, 15):Month(1):Date(2020, 3, 15)
# dimensional information:
dimensions = (Lon(lons), Lat(lats), Time(t))
data = rand(36, 37, 241) # numeric data
A = ClimArray(data, dimensions)

ClimateBase.gnv — Function

gnv(object) → x

Short for "get numeric value", this function will return the pure numeric value(s) of the given object. Convenience function for quickly getting the numeric data of dimensional arrays or dimensions.

It is strongly recommended to use the dimensions we export (because we dispatch on them and use their information):

for D in ClimateBase.STANDARD_DIMS
    println(D, " (full name = $(DimensionalData.name(D)))")
end

Lon (full name = Longitude)
Lat (full name = Latitude)
Ti (full name = Time)
Hei (full name = Height)
Pre (full name = Pressure)
Coord (full name = )

We explicitly assume that Lon, Lat are measured in degrees and not radians or meters (extremely important for spatial averaging processes).

Crash-course to DimensionalData.jl

DimensionalData — Module

DimensionalData

DimensionalData.jl provides tools and abstractions for working with datasets that have named dimensions, and optionally a lookup index. It provides no-cost abstractions for named indexing, and fast index lookups.

DimensionalData is a pluggable, generalised version of AxisArrays.jl with a cleaner syntax, and additional functionality found in NamedDims.jl. It has similar goals to pythons xarray, and is primarily written for use with spatial data in Rasters.jl.

The basic syntax is:

julia> using DimensionalData

julia> rand(X(10.0:40.0), Y(50))
31×50 DimArray{Float64,2} with dimensions:
  X Sampled 10.0:1.0:40.0 ForwardOrdered Regular Points,
  Y
 0.793097  0.489866  0.462396  …  0.910434  0.850573   0.183605
 0.76277   0.737544  0.290279     0.742267  0.686086   0.530159
 ⋮                             ⋱
 0.281043  0.979182  0.868658     0.642477  0.139536   0.540512
 0.546036  0.83382   0.530098  …  0.351608  0.0385814  0.159299

julia> A[Y=1, X=1:10]
10-element DimArray{Float64,1} with dimensions:
  X Sampled 10.0:1.0:19.0 ForwardOrdered Regular Points
and reference dimensions: Y
 0.245691
 0.902444
 0.777441
 ⋮
 0.744612
 0.440409
 0.631956

See the docs for more details

Some properties of DimensionalData.jl objects:

broadcasting and most Base methods maintain and sync dimension context.

comprehensive plot recipes for Plots.jl.

a Tables.jl interface with DimTable

multi-layered DimStacks that can be indexed together, and have base methods applied to all layers.

the Adapt.jl interface for use on GPUs, even as GPU kernel arguments.

traits for handling a wide range of spatial data types accurately.

Methods where dims can be used containing indices or Selectors

getindex, setindex! view

Methods where dims, dim types, or Symbols can be used to indicate the array dimension:

size, axes, firstindex, lastindex

cat, reverse, dropdims

reduce, mapreduce

sum, prod, maximum, minimum,

mean, median, extrema, std, var, cor, cov

permutedims, adjoint, transpose, Transpose

mapslices, eachslice

Methods where dims can be used to construct DimArrays:

fill, ones, zeros, falses, trues, rand

Note: recent changes have greatly reduced the exported API

Previously exported methods can me brought into global scope by using the sub-modules they have been moved to - LookupArrays and Dimensions:

using DimensionalData
using DimensionalData.LookupArrays, DimensionalData.Dimensions

Alternate Packages

There are a lot of similar Julia packages in this space. AxisArrays.jl, NamedDims.jl, NamedArrays.jl are registered alternative that each cover some of the functionality provided by DimensionalData.jl. DimensionalData.jl should be able to replicate most of their syntax and functionality.

AxisKeys.jl and AbstractIndices.jl are some other interesting developments. For more detail on why there are so many similar options and where things are headed, read this thread.

The main functionality is explained here, but the full list of features is listed at the API page.

Available selectors

Selector	Description

[At(x)]	get the index exactly matching the passed in value(s)

[Near(x)]	get the closest index to the passed in value(s)

[Contains(x)]	get indices where the value x falls within an interval

[Where(f)]	filter the array axis by a function of the dimension index values.

[a..b]	get all indices between two values, inclusively.

[OpenInterval(a, b)]	get all indices between a and b, exclusively.

[Interval{A,B}(a, b)]	get all indices between a and b, as :closed or :open.